49: Make "streaming" subproject buildable (#60)

+++ streaming/streaming_common/decoder.cpp
@@ -141,7 +141,7 @@ bool Decoder::upload() {
                                    &packet_->data,
                                    &packet_->size,
                                    buffer_.data(),
-                                   size,
+                                   static_cast<int>(size),
                                    AV_NOPTS_VALUE,
                                    AV_NOPTS_VALUE,
                                    0);
